feat: LEAP-355: Enhance Relation helpers and fix ts typings (#23)
* feat: LEAP-355: Enhance Relation helpers and fix ts typings

* Enhance ToolBar with additional helper methods

Added more detailed commentary to the existing functions in ToolBar.ts. Furthermore, enhanced the functionalities by introducing new functions for interacting with annotation elements. These new features and methods will help in interacting with the annotations dropdown, creating new annotations, and selecting specific annotations from the list.

* Rename Relation methods for clarity

The 'relation' function was renamed to 'getRelation', and the hover/unhover functions were renamed to 'hoverOverRelation' and 'stopHoveringOverRelation'. This was done to improve readability and better express the functionalities of these methods in the LSF Relations helper file.

* Add blur event after filing relation lebel input